SmartvattenAPI/requirements.txt
argcomplete
influxdb
pandas
python-dateutil
pytz
requests
